fix lỗi deploy

+++ tests/db-handler.js
@@ -10,14 +10,29 @@ module.exports.connect = async () => {
 };
 
 module.exports.closeDatabase = async () => {
-    await mongoose.connection.dropDatabase();
-    await mongoose.connection.close();
-    if (mongoServer) {
-        await mongoServer.stop();
+    try {
+        if (mongoose.connection.readyState === 1) {
+            await mongoose.connection.dropDatabase();
+        }
+
+        if (mongoose.connection.readyState !== 0) {
+            await mongoose.connection.close(true);
+        }
+
+        if (mongoServer) {
+            await mongoServer.stop();
+            mongoServer = undefined;
+        }
+    } catch (error) {
+        // Teardown should be safe to call more than once.
     }
 };
 
 module.exports.clearDatabase = async () => {
+    if (mongoose.connection.readyState !== 1) {
+        return;
+    }
+
     const { collections } = mongoose.connection;
     for (const key in collections) {
         const collection = collections[key];
